[WLD-133] CLITester helper moved to a separate package (can be used as a dependency)
